Crate profiles_integrity

profiles_integrity

Profiles zome for any Holochain app.

If you need to manage profiles (nickname, name, avatar, age and other useful personal information) you can directly include this zome and its coordinator counterpart in your DNA.

Read about how to include both this zome and its frontend module in your application here.

Structs

Profile

Profile entry definition.

Enums

EntryTypes

LinkTypes

UnitEntryTypes

Functions

__num_entry_types

__num_link_types

action_hash

entry_defs

genesis_self_check

validate

validate_agent_joining

validate_create_profile

validate_delete_profile

validate_update_profile